+++ src/Serwisant/SerwisantApi/Types/SchemaService/Repair.php
@@ -73,6 +73,11 @@ class Repair extends Types\Type
   public $deliveryType;
 
   /**
+   * @var Device
+  */
+  public $device;
+
+  /**
    * @var RepairDiagnosis
   */
   public $diagnosis;