+++ src/Serwisant/SerwisantApi/Types/SchemaService/ServiceQuery.php
@@ -63,6 +63,18 @@ class ServiceQuery extends Types\RootType
   }
 
   /**
+   * @param int $limit
+   * @param int $page
+   * @param DevicesFilter $filter
+   * @param string $sort
+   * @return DevicesResult
+   */
+  public function devices(int $limit = null, int $page = null, DevicesFilter $filter = null, string $sort = null, $vars = array())
+  {
+     return $this->inputArgs('devices', array_merge($vars, ['limit' => $limit, 'page' => $page, 'filter' => $filter, 'sort' => $sort]));
+  }
+
+  /**
    * Return entries for specified dictionary type
    * @param DictionaryEntriesFilter $filter
    * @return Dictionary[]